+++ zcash_primitives/src/lib.rs
@@ -18,7 +18,6 @@ pub mod note_encryption;
 pub mod pedersen_hash;
 pub mod primitives;
 pub mod prover;
-pub mod redjubjub;
 pub mod sapling;
 pub mod serialize;
 pub mod transaction;
